rebase

+++ Chalkable.Web/Controllers/AnnouncementAttachmentController.cs
@@ -5,6 +5,7 @@ using Chalkable.Common;
 using Chalkable.Common.Exceptions;
 using Chalkable.Common.Web;
 using Chalkable.Web.ActionFilters;
+using Chalkable.Web.ActionResults;
 using Chalkable.Web.Logic;
 using Chalkable.Web.Models;
 using Chalkable.Web.Models.AnnouncementsViewData;
@@ -21,33 +22,75 @@ namespace Chalkable.Web.Controllers
         [AcceptVerbs(HttpVerbs.Post), AuthorizationFilter("SysAdmin, AdminGrade, AdminEdit, AdminView, Teacher, Student")]
         public ActionResult AddAttachment(int announcementId)
         {
-            byte[] bin;
-            string name;
-            if (!GetFileFromRequest(out bin, out name))
+            try
             {
-                return Json(new ChalkableException(ChlkResources.ERR_FILE_REQUIRED));
-            }
-            string uuid = null;
-            if (SchoolLocator.CrocodocService.IsDocument(name))
-            {
-                try
+                EnsureAnnouncementExsists(announcementId);            
+
+                byte[] bin;
+                string name;
+                if (!GetFileFromRequest(out bin, out name))
+                {
+                    return Json(new ChalkableException(ChlkResources.ERR_FILE_REQUIRED));
+                }
+                string uuid = null;
+                if (SchoolLocator.CrocodocService.IsDocument(name))
                 {
                     uuid = SchoolLocator.CrocodocService.UploadDocument(name, bin).uuid;
                 }
-                catch (ChalkableException exception)
-                {
-                    return Json(exception);
-                }
+                var announcement = SchoolLocator.AnnouncementAttachmentService.AddAttachment(announcementId, bin, name, uuid);
+                AnnouncementViewData res = PrepareFullAnnouncementViewData(announcement.Id);
+                return Json(res, HTML_CONTENT_TYPE, 6);
             }
-            var announcement = SchoolLocator.AnnouncementAttachmentService.AddAttachment(announcementId, bin, name, uuid);
-            AnnouncementViewData res = PrepareFullAnnouncementViewData(announcement.Id);
-            return Json(res, HTML_CONTENT_TYPE, 6);
+            catch (Exception exception)
+            {
+                Response.StatusCode = 500;
+                return new ChalkableJsonResult(false)
+                {
+                    Data = ExceptionViewData.Create(exception, exception.InnerException),
+                    ContentType = HTML_CONTENT_TYPE,
+                    SerializationDepth = 4
+                };
+            }
+        }
+
+        private void EnsureAnnouncementExsists(int announcementId)
+        {
+            // get announcement to ensure it exists
+            SchoolLocator.AnnouncementService.GetAnnouncementById(announcementId);
+        }
+
+        [AuthorizationFilter("SysAdmin, AdminGrade, AdminEdit, AdminView, Teacher, Student")]
+        public ActionResult CloneAttachment(int originalAttachmentId, int announcementId)
+        {
+            EnsureAnnouncementExsists(announcementId);
+
+            var attContentInfo = SchoolLocator.AnnouncementAttachmentService.GetAttachmentContent(originalAttachmentId);
+            if (attContentInfo != null && announcementId == attContentInfo.Attachment.AnnouncementRef)
+            {
+                byte[] bin = attContentInfo.Content;
+                string name = attContentInfo.Attachment.Name;
+                string uuid = null;
+                if (SchoolLocator.CrocodocService.IsDocument(name))
+                {
+                    uuid = SchoolLocator.CrocodocService.UploadDocument(name, bin).uuid;
+                }
+                
+                SchoolLocator.AnnouncementAttachmentService.AddAttachment(attContentInfo.Attachment.AnnouncementRef, bin, name, uuid);
+            }
+            AnnouncementViewData res = PrepareFullAnnouncementViewData(announcementId);
+            return Json(res, 6);
         }
 
         [AuthorizationFilter("SysAdmin, AdminGrade, AdminEdit, AdminView, Teacher, Student")]
         public ActionResult DownloadAttachment(int announcementAttachmentId, bool? needsDownload, int? width, int? height)
         {
             var attContentInfo = SchoolLocator.AnnouncementAttachmentService.GetAttachmentContent(announcementAttachmentId);
+            if (attContentInfo == null)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+
             var attName = attContentInfo.Attachment.Name;
             var content = attContentInfo.Content;
             var contentTypeName = MimeHelper.GetContentTypeByName(attName);
@@ -64,16 +107,22 @@ namespace Chalkable.Web.Controllers
         }
 
         [AuthorizationFilter("SysAdmin, AdminGrade, AdminEdit, AdminView, Teacher, Student")]
-        public ActionResult DeleteAttachment(int announcementAttachmentId)
+        public ActionResult DeleteAttachment(int announcementAttachmentId, int announcementId)
         {
+            EnsureAnnouncementExsists(announcementId);
+
             var attachment = SchoolLocator.AnnouncementAttachmentService.GetAttachmentById(announcementAttachmentId);
-            SchoolLocator.AnnouncementAttachmentService.DeleteAttachment(announcementAttachmentId);
-            var res = PrepareFullAnnouncementViewData(attachment.AnnouncementRef);
+            if (attachment != null && attachment.AnnouncementRef == announcementId)
+                SchoolLocator.AnnouncementAttachmentService.DeleteAttachment(announcementAttachmentId);
+
+            var res = PrepareFullAnnouncementViewData(announcementId);
             return Json(res, 6);
         }
         [AuthorizationFilter("SysAdmin, AdminGrade, AdminEdit, AdminView, Teacher, Student")]
         public ActionResult GetAttachments(int announcementId, int? start, int? count)
         {
+            EnsureAnnouncementExsists(announcementId);
+
             var announcementAttachments = SchoolLocator.AnnouncementAttachmentService.GetAttachments(announcementId, start ?? 0, count ?? 10, false);
             var attachmentsInfo = AttachmentLogic.PrepareAttachmentsInfo(announcementAttachments);
             var res = AnnouncementAttachmentViewData.Create(attachmentsInfo, SchoolLocator.Context.PersonId ?? 0);
@@ -84,6 +133,12 @@ namespace Chalkable.Web.Controllers
         public ActionResult StartViewSession(int announcementAttachmentId)
         {
             var att = SchoolLocator.AnnouncementAttachmentService.GetAttachmentById(announcementAttachmentId);
+            if (att == null)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+
             try
             {
                 var person = SchoolLocator.PersonService.GetPerson(SchoolLocator.Context.PersonId ?? 0);
@@ -105,9 +160,15 @@ namespace Chalkable.Web.Controllers
                     });
                 return Json(res.session);
             }
-            catch (Exception ex)
+            catch (Exception exception)
             {
-                return Json(new ChalkableException(ex.Message, ex));
+                Response.StatusCode = 500;
+                return new ChalkableJsonResult(false)
+                {
+                    Data = ExceptionViewData.Create(exception, exception.InnerException),
+                    ContentType = HTML_CONTENT_TYPE,
+                    SerializationDepth = 4
+                };
             }
         }
